requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
stripe==12.0.1
gunicorn==23.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
flask-wtf==1.2.2
WTForms==3.2.1
flask_babel==4.0.0
pymysql==1.1.1
python-dotenv==1.1.0
twilio==9.5.2
